[MERGE_UNSURE] use bytes for vertex info
further reduces vram usage: down to 8MB in test world.
Huge odd decrease in performance on the Radeon HD6850, unsure to merge this into master

+++ include/chunkmesher.hpp
@@ -17,9 +17,9 @@ namespace chunkmesher{
 	Chunk::Chunk* chunk; 
 	GLuint numVertices{0};
 
-	std::vector<GLfloat> vertices;
-	std::vector<GLfloat> extents;
-	std::vector<GLfloat> texinfo;
+	std::vector<GLubyte> vertices;
+	std::vector<GLubyte> extents;
+	std::vector<GLubyte> texinfo;
     };
     oneapi::tbb::concurrent_queue<MeshData*>& getMeshDataQueue();
